Fix attendance not showing properly in report card Change type of days in Attendance table to all be Float

+++ database/migrations/2018_07_07_022948_create_attendance_record.php
@@ -20,10 +20,10 @@ class CreateAttendanceRecord extends Migration
            $table->dateTimeTz('datetime');
             $table->timestamps();
 
-           $table->unsignedTinyInteger('late');
-           $table->unsignedTinyInteger('absent');
-           $table->unsignedTinyInteger('leave');
-           $table->unsignedTinyInteger('sick');
+           $table->float('late');
+           $table->float('absent');
+           $table->float('leave');
+           $table->float('sick');
            $table->unsignedTinyInteger('data_status');
 
            $table->primary(['student_id','semester','academic_year','datetime'],'attendance_record_primary');
